Fix post use likeability trait
